fix: move getLineOffsetYBinding and getLineOffsetXBinding functions into BindingsUtils
