+++ graphedit/src/main/java/dk/gtz/graphedit/syntaxes/lts/view/TransitionController.java
@@ -9,10 +9,9 @@ import dk.gtz.graphedit.syntaxes.lts.viewmodel.ViewModelTransition;
 import dk.gtz.graphedit.tool.ITool;
 import dk.gtz.graphedit.view.EdgeController;
 import dk.gtz.graphedit.view.ISyntaxFactory;
+import dk.gtz.graphedit.view.util.BindingsUtil;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.gtz.graphedit.viewmodel.ViewModelProjectResource;
-import javafx.beans.binding.Bindings;
-import javafx.beans.binding.DoubleBinding;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleDoubleProperty;
@@ -37,36 +36,15 @@ public class TransitionController extends EdgeController {
     private Label createActionLabel() {
         var result = new Label(edge.action().get());
         result.getStyleClass().add("outline");
-        result.translateXProperty().bind(getLineOffsetXBinding(
+        result.translateXProperty().bind(BindingsUtil.getLineOffsetXBinding(
                     line.startXProperty(), line.startYProperty(),
                     line.endXProperty(), line.endYProperty(),
                     labelDirOffset).subtract(result.widthProperty().divide(2)));
-        result.translateYProperty().bind(getLineOffsetYBinding(
+        result.translateYProperty().bind(BindingsUtil.getLineOffsetYBinding(
                     line.startXProperty(), line.startYProperty(),
                     line.endXProperty(), line.endYProperty(),
                     labelDirOffset).subtract(result.heightProperty().divide(2)));
         return result;
-    }
-
-    // TODO: Move these into BindingsUtil and add javadocs
-    private DoubleBinding getLineOffsetXBinding(DoubleProperty lineStartX, DoubleProperty lineStartY, DoubleProperty lineEndX, DoubleProperty lineEndY, DoubleProperty scalar) {
-        return Bindings.createDoubleBinding(() -> {
-            var dirX = lineEndX.get() - lineStartX.get();
-            var dirY = lineEndY.get() - lineStartY.get();
-            var len = Math.sqrt(Math.pow(dirX, 2) + Math.pow(dirY, 2));
-            var nDirX = dirX / len;
-            return lineStartX.get() + (nDirX * (scalar.get() * len));
-        }, lineStartX, lineStartY, lineEndX, lineEndY, scalar);
-    }
-
-    private DoubleBinding getLineOffsetYBinding(DoubleProperty lineStartX, DoubleProperty lineStartY, DoubleProperty lineEndX, DoubleProperty lineEndY, DoubleProperty scalar) {
-        return Bindings.createDoubleBinding(() -> {
-            var dirX = lineEndX.get() - lineStartX.get();
-            var dirY = lineEndY.get() - lineStartY.get();
-            var len = Math.sqrt(Math.pow(dirX, 2) + Math.pow(dirY, 2));
-            var nDirY = dirY / len;
-            return lineStartY.get() + (nDirY * (scalar.get() * len));
-        }, lineStartX, lineStartY, lineEndX, lineEndY, scalar);
     }
 }
